Revert "超市管理系统1.0"
This reverts commit f4feb88821fd1494c277701eef35e238d4511027.

# Conflicts:
#	src/main/java/cn/smbms/controller/BillController.java
#	src/main/java/cn/smbms/controller/ProviderController.java
#	src/main/java/cn/smbms/controller/UserController.java
#	src/main/java/cn/smbms/dao/BillMapper.java
#	src/main/java/cn/smbms/dao/BillMapper.xml
#	src/main/java/cn/smbms/service/BillServiceImpl.java
#	src/main/java/cn/smbms/service/ProviderService.java
#	src/main/java/cn/smbms/service/ProviderServiceImpl.java
#	src/main/java/cn/smbms/service/UserService.java
#	src/main/java/cn/smbms/service/UserServiceImpl.java
#	src/main/webapp/WEB-INF/jsp/billlist.jsp
#	src/main/webapp/WEB-INF/jsp/billmodify.jsp
#	src/main/webapp/WEB-INF/jsp/common/head.jsp
#	src/main/webapp/WEB-INF/jsp/providerlist.jsp
#	src/main/webapp/WEB-INF/jsp/pwdmodify.jsp
#	src/main/webapp/static/js/billlist.js
#	src/main/webapp/static/js/pwdmodify.js
#	target/classes/cn/smbms/controller/BillController.class
#	target/classes/cn/smbms/controller/ProviderController.class
#	target/classes/cn/smbms/controller/UserController.class
#	target/classes/cn/smbms/dao/BillMapper.class
#	target/classes/cn/smbms/dao/BillMapper.xml
#	target/classes/cn/smbms/service/BillServiceImpl.class
#	target/classes/cn/smbms/service/ProviderService.class
#	target/classes/cn/smbms/service/ProviderServiceImpl.class
#	target/classes/cn/smbms/service/UserService.class
#	target/classes/cn/smbms/service/UserServiceImpl.class
#	target/supermarket/WEB-INF/classes/cn/smbms/controller/BillController.class
#	target/supermarket/WEB-INF/classes/cn/smbms/controller/ProviderController.class
#	target/supermarket/WEB-INF/classes/cn/smbms/controller/UserController.class
#	target/supermarket/WEB-INF/classes/cn/smbms/dao/BillMapper.class
#	target/supermarket/WEB-INF/classes/cn/smbms/dao/BillMapper.xml
#	target/supermarket/WEB-INF/classes/cn/smbms/service/BillServiceImpl.class
#	target/supermarket/WEB-INF/classes/cn/smbms/service/ProviderService.class
#	target/supermarket/WEB-INF/classes/cn/smbms/service/ProviderServiceImpl.class
#	target/supermarket/WEB-INF/classes/cn/smbms/service/UserService.class
#	target/supermarket/WEB-INF/classes/cn/smbms/service/UserServiceImpl.class
#	target/supermarket/WEB-INF/jsp/billlist.jsp
#	target/supermarket/WEB-INF/jsp/billmodify.jsp
#	target/supermarket/WEB-INF/jsp/common/head.jsp
#	target/supermarket/WEB-INF/jsp/providerlist.jsp
#	target/supermarket/WEB-INF/jsp/pwdmodify.jsp
#	target/supermarket/static/js/billlist.js
#	target/supermarket/static/js/pwdmodify.js

+++ src/main/java/cn/smbms/pojo/Bill.java
@@ -28,9 +28,9 @@ public class Bill {
 
     private Date modifyDate;
 
-    private Long providerId;
+    private Integer providerId;
 
-    public Bill(Long id, String billCode, String productName, String productDesc, String productUnit, BigDecimal productCount, BigDecimal totalPrice, Integer isPayment, Long createdBy, Date creationDate, Long modifyBy, Date modifyDate, Long providerId) {
+    public Bill(Long id, String billCode, String productName, String productDesc, String productUnit, BigDecimal productCount, BigDecimal totalPrice, Integer isPayment, Long createdBy, Date creationDate, Long modifyBy, Date modifyDate, Integer providerId) {
         this.id = id;
         this.billCode = billCode;
         this.productName = productName;
@@ -146,11 +146,11 @@ public class Bill {
         this.modifyDate = modifyDate;
     }
 
-    public Long getProviderId() {
+    public Integer getProviderId() {
         return providerId;
     }
 
-    public void setProviderId(Long providerId) {
+    public void setProviderId(Integer providerId) {
         this.providerId = providerId;
     }
 }
